Maybe it will fix issue 24 attempt 2
