+++ .idea/deploymentTargetSelector.xml
@@ -4,10 +4,10 @@
     <selectionStates>
       <SelectionState runConfigName="app">
         <option name="selectionMode" value="DROPDOWN" />
-        <DropdownSelection timestamp="2024-09-13T08:08:25.022226Z">
+        <DropdownSelection timestamp="2025-02-20T23:21:09.100243100Z">
           <Target type="DEFAULT_BOOT">
             <handle>
-              <DeviceId pluginId="PhysicalDevice" identifier="serial=2B952BE1" />
+              <DeviceId pluginId="LocalEmulator" identifier="path=/home/wiktor/.android/avd/Medium_Tablet_API_29.avd" />
             </handle>
           </Target>
         </DropdownSelection>
